Actually use the C11 atomic API.
The alternative definitions in utils.h do not work here, because in
C11 atomic_compare_exchange(p,q,v) means

	if (*p == *q)
		*p = v;
	else
		*q = *p;

That is, the second argument is a _pointer_ to the expected value,
which is replaced by the actual value in the event of failure.

Note: This admits a slight simplification in the loops -- we can
avoid repeating atomic_load_relaxed because we already have the
modified value in the event of a race.